Handle no event raffle case in context of Raffle API
Building on commit 39e3acfdb1f7, we need to make sure Raffle API context
handles requests for a raffle with no provided events in a proper way.

RaffleApiContext implements the step for a scenario when a raffle is
requested, but no events are provided.

RaffleApiController returns appropriate message in such cases.