Refactor JoindInTalk, Raffle&RaffleTest
Raffle had nested looping over events and talks just to get to comments,
and was hard to test with a lot of mocking.

Removed one level of nested looping by introducing a comment counter on
JoindInTalk and refactored the Raffle a bit.

Rewrote some of the Raffle tests and removed some others that didn't make
sense.

Most of the work done by @msvrtan.