Add scenario for accessing noneligible comments
As users, we should have access to noneligible comments as well as to
eligible ones.

This commit will add behat scenarios for accessing noneligible comments
through both the application and API.

JoindInUser Entity had to be slightly modified, so it could create it's
own Collection for storing comments at the instatiation.